Create actions to create notes for events, move note actions to notes.js
