+++ src/actions/types.js
@@ -22,3 +22,7 @@ export const GET_NOTES_BY_EVENT = 'get_notes_by_event';
 export const WRITE_EVENT = 'write_event';
 export const ADD_EVENT = 'add_event';
 export const DELETE_EVENT = 'delete_event';
+
+export const WRITE_NOTE = 'write_note';
+export const CLEAR_NOTE = 'clear_note';
+export const ADD_EVENT_NOTE = 'add_note';
